requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
feedparser>=6.0.0
python-dateutil==2.9.0
pytz==2025.1